return time as long

+++ src/org/dinopolis/gpstool/gpsinput/garmin/GarminTrackpoint.java
@@ -66,7 +66,7 @@ public interface GarminTrackpoint
  * @return Time (seconds)
  * @throws GarminUnsupportedMethodException
  */
- public float getTime() throws GarminUnsupportedMethodException;
+ public long getTime() throws GarminUnsupportedMethodException;
 	
 //----------------------------------------------------------------------
 /**
